+++ src/test/java/com/remake/gone/auth/service/PhoneAuthServiceTest.java
@@ -14,10 +14,10 @@ import com.remake.gone.auth.dto.PhoneSendCodeResponse;
 import com.remake.gone.auth.dto.PhoneVerifyCodeRequest;
 import com.remake.gone.auth.dto.PhoneVerifyCodeResponse;
 import com.remake.gone.auth.exception.AuthErrorCode;
-import com.remake.gone.auth.utils.SmsSender;
 import com.remake.gone.common.exception.CustomException;
 import com.remake.gone.common.redis.RedisKeyType;
 import com.remake.gone.common.redis.RedisRepository;
+import com.remake.gone.sms.SmsSender;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
